completed design refactoring
